Basic suppport for fetching Nighlies
Don't know how to search in PullRequest via v4 API within
single repostiory. So this commit uses a workaoround which
can fetch more than GitHub API limit (100) of result, filter
them locally a fetch detailed data.

It still truncates the resulting to 100.

The tests are bit unstable and seems to slow down and when they are
slow they tend to affect each other. Would be good to mock them.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>