Fix CSS build when project is in some "src" dir
If whole project was located in a directory with "src" in path,
CSS build failed. Now it succeeds.

Original code was structured that way to work on Windows and Linux
but given the fact that no variables are use we can remove the
usage and still be functional.

Fixes Issue #1

Signed-off-by: Petr Vobornik <pvoborni@gmail.com>